update aula 03A parte final

+++ demo/src/main/java/app/controller/CarroController.java
@@ -36,7 +36,7 @@ public class CarroController {
 	
 	// método para buscar um carro pelo id
 	@GetMapping("/findById/{id}") // 
-	public ResponseEntity<Carro> findById(@PathVariable int id) {
+	public ResponseEntity<Carro> findById(@PathVariable Long id) {
 		
 		try {
 			Carro carro = this.carroService.findById(id);
